Combine Mangal networks

Combine mgNetworksCollection and mgNetwork objects into a mgNetworksCollection object.

combine_mgNetworks(...)

Arguments

...
objects of class mgNetworksCollection or mgNetwork or a list #' of objects of these classes.

Value

An object of class mgNetworksCollection

Examples

mg_random_1071 <- get_collection(c(1071), verbose = FALSE)

#> Processing id: 75 	1/1  

#>                                                                              

mg_random_1074 <- get_collection(c(1074), verbose = FALSE)

#> Processing id: 75 	1/1  

#>                                                                              

combine_mgNetworks(mg_random_1071, mg_random_1074)

#> A collection of 2 networks
#> 
#> * Network # from data set #
#> * Description: structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Bajo
#> * Includes 4 edges and 6 nodes 
#> * Current taxonomic IDs coverage for nodes of this network: 
#>   --> ITIS: 100%, BOLD: 83%, EOL: 100%, COL: 100%, GBIF: 100%, NCBI: 100%
#> * Published in ref # DOI:10.1007/s10641-010-9606-0
#> 
#> * Network # from data set #
#> * Description: structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Fukui
#> * Includes 6 edges and 7 nodes 
#> * Current taxonomic IDs coverage for nodes of this network: 
#>   --> ITIS: 100%, BOLD: 86%, EOL: 100%, COL: 100%, GBIF: 100%, NCBI: 100%
#> * Published in ref # DOI:10.1007/s10641-010-9606-0
#>