Get a collection of networks

Retrieve a set of networks based on the results of one of the search_*() function. The function also accept a numeric vector of Mangal network IDs.

get_collection(x, ...)

# S3 method for default
get_collection(x, ...)

# S3 method for mgSearchDatasets
get_collection(x, ...)

# S3 method for mgSearchNetworks
get_collection(x, ...)

# S3 method for mgSearchReferences
get_collection(x, ...)

# S3 method for mgSearchNodes
get_collection(x, ...)

# S3 method for mgSearchTaxonomy
get_collection(x, ...)

# S3 method for mgSearchInteractions
get_collection(x, ...)

Arguments

x
numeric vector of Mangal network IDs or an object returned by by one of the search_*() functions.

...
arguments to be passed on to rmangal::get_network_by_id().

Value

If there is only one network to be retrieved, get_collection() returns a mgNetwork object, otherwise it returns a object of class mgNetworksCollection which is a collection (a list) of mgNetwork objects rmangal::get_network_by_id()).

Methods (by class)

default: Get a collection of networks (default).

mgSearchDatasets: Get a collection of networks from a mgSearchDatasets object.

mgSearchNetworks: Get a collection of networks from a mgSearchNetworks object.

mgSearchReferences: Get a collection of networks from a mgSearchReferences object.

mgSearchNodes: Get a collection of networks from a mgSearchNodes object.

mgSearchTaxonomy: Get a collection of networks from a mgSearchTaxa object.

mgSearchInteractions: Get a collection of networks from a mgSearchTaxa object.

See also

Examples

# \donttest{
mg_2 <- get_collection(c(1076:1077), verbose = FALSE)

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

mg_anemone <- get_collection(search_networks(query='anemone%'), verbose = FALSE)

#> Data retrieval 100%   

#>                                                                              

#> Found 16 networks

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

#> Processing id: 75 	1/1  

#>                                                                              

# }