Generic API function to retrieve singletons

Generic API function to retrieve singletons

get_singletons(endpoint = NULL, ids = NULL, output = "list",
  flatten = TRUE, ...)

Arguments

endpoint
character API entry point

ids
numeric vector of ids

output
character output type (data.frame, list, spatial, raw) return; default: list

flatten
logical flatten nested data.frame, see jsonlite::flatten(); default: TRUE

...
httr options, see httr::GET()

Value

Object of class mgGetResponses whithin each level is the specific ID called. Each item of the mgGetResponses list corresponds to an API call. Each call returns an object:

getSuccess which is a list with the body httr::content() and the server response httr::response().

getError which has the exact same structure with an empty body.

Details

See endpoints available with endpoints()