Search mangal datasets

Search mangal datasets

search_datasets(query = NULL, ...)

Arguments

query
character keyword used to search (case sensitive), or list of custom query.

...
arguments from rmangal::get_gen()

Value

object data.frame: Datasets with all networks and the original reference attached.

Examples

# NOT RUN {
all_datasets <- search_datasets()
all_datasets
class(all_datasets)
# }
# Search with keyword
search_datasets(query = "lagoon")

#> Found 2 dataset(s) for query: lagoon

#> $id
#> [1] 22 52
#> 
#> $name
#> [1] "zetina_2003" "yanez_1978" 
#> 
#> $date
#> [1] "2003-01-01T05:00:00.000Z" "1973-09-01T04:00:00.000Z"
#> 
#> $description
#> [1] "Dietary matrix of the Huizache–Caimanero lagoon"
#> [2] "Food web of the Guerrero lagoon"                
#> 
#> $public
#> [1] TRUE TRUE
#> 
#> $created_at
#> [1] "2019-02-23T17:04:32.017Z" "2019-02-24T23:42:52.062Z"
#> 
#> $updated_at
#> [1] "2019-02-23T17:04:32.017Z" "2019-02-24T23:42:52.062Z"
#> 
#> $ref_id
#> [1] 22 53
#> 
#> $user_id
#> [1] 3 3
#> 
#> $networks
#> $networks[[1]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1    86 zeti… 2003… Dietary ma… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[2]]
#> # A tibble: 2 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1   927 yane… 1973… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 2   926 yane… 1973… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> 
#> $references
#> $references[[1]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    22 10.1… manue… 2003  NA    NA    "@art… https://… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[2]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    53 NA    yanez… 1978  NA    NA    @book… http://b… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> 
#> attr(,"row.names")
#> [1] 1 2
#> attr(,"class")
#> [1] "mangalSearch"

search_datasets(query = "2011")

#> Found 8 dataset(s) for query: 2011

#> $id
#> [1]  32  70  73  74  79  94  97 155
#> 
#> $name
#> [1] "gontikaki_2011"        "kohler_2011"           "filgueira_2011"       
#> [4] "baeta_2011"            "douglas_2011"          "fetahi_et_al_2011"    
#> [7] "hampton_et_al_2011"    "filgueira_castro_2011"
#> 
#> $date
#> [1] "2007-05-01T04:00:00.000Z" "2011-01-01T05:00:00.000Z"
#> [3] "2000-02-01T05:00:00.000Z" "1993-01-01T05:00:00.000Z"
#> [5] "2011-01-01T05:00:00.000Z" "2007-10-01T04:00:00.000Z"
#> [7] "2004-06-01T04:00:00.000Z" "2000-02-01T05:00:00.000Z"
#> 
#> $description
#> [1] "Food web of a sub-arctic deep-sea benthic community"                                              
#> [2] "Hummingbirds-flowers interactions in an altitudinal gradient in the Brazilian Atlantic Rainforest"
#> [3] "Food web of the Ria de Vigo"                                                                      
#> [4] "Food web of the Mondego bay estuary"                                                              
#> [5] "Food web of the Chesapeake's Bay Eelgrass Bed"                                                    
#> [6] "Food web structure and diet composition based on an Ecopath model in Lake Hayq, Ethiopia"         
#> [7] "Food web structure of a deep oligotrophic lake (Lake Crescent), Washington, USA"                  
#> [8] "Food web structure of an intertidal zone in San Simon Bay, Spain"                                 
#> 
#> $public
#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> $created_at
#> [1] "2019-02-24T15:16:24.404Z" "2019-02-25T21:28:48.689Z"
#> [3] "2019-02-26T19:15:58.087Z" "2019-02-26T19:50:23.243Z"
#> [5] "2019-02-27T00:20:35.186Z" "2019-03-01T20:10:53.367Z"
#> [7] "2019-03-04T15:45:31.613Z" "2019-03-15T16:14:00.003Z"
#> 
#> $updated_at
#> [1] "2019-02-24T15:16:24.404Z" "2019-02-25T21:28:48.689Z"
#> [3] "2019-02-26T19:15:58.087Z" "2019-02-26T19:50:23.243Z"
#> [5] "2019-02-27T00:20:35.186Z" "2019-03-01T20:10:53.367Z"
#> [7] "2019-03-04T15:45:31.613Z" "2019-03-15T16:14:00.003Z"
#> 
#> $ref_id
#> [1]  32  69  72  73  78  93  96 152
#> 
#> $user_id
#> [1] 3 2 3 3 3 4 4 4
#> 
#> $networks
#> $networks[[1]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1    96 gont… 2007… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[2]]
#> # A tibble: 3 x 11
#>      id name  date  geom  description public all_interactions created_at
#>   <int> <chr> <chr> <lgl> <chr>       <lgl>  <lgl>            <chr>     
#> 1   954 kohl… 2011… NA    Hummingbir… TRUE   FALSE            2019-02-2…
#> 2   953 kohl… 2011… NA    Hummingbir… TRUE   FALSE            2019-02-2…
#> 3   952 kohl… 2011… NA    Hummingbir… TRUE   FALSE            2019-02-2…
#> # … with 3 more variables: updated_at <chr>, dataset_id <int>, user_id <int>
#> 
#> $networks[[3]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1115 filg… 2000… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[4]]
#> # A tibble: 6 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1065 baet… 1999… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 2  1066 baet… 2001… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 3  1064 baet… 1993… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 4  1069 baet… 2001… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 5  1068 baet… 1999… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 6  1067 baet… 1993… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[5]]
#> # A tibble: 3 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1113 doug… 2006… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 2  1114 doug… 2006… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 3  1112 doug… 2005… Food web o… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[6]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1464 feta… 2007… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[7]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1469 hamp… 2004… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> $networks[[8]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1  1567 filg… 2000… Food web s… TRUE   FALSE            2019-03-1… 2019-03-1…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> 
#> $references
#> $references[[1]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    32 10.1… e. go… 2011  NA    NA    @arti… https://… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[2]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    69 NA    kohler 2011  NA    NA    NA     null      https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[3]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    72 10.1… r. fi… 2011  NA    NA    @arti… https://… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[4]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    73 10.1… alexa… 2011  NA    NA    @arti… https://… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[5]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    78 10.1… james… 2011  NA    NA    @arti… https://… https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[6]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    93 10.1… fetahi 2011  NA    NA    @arti… https://… https:/… 2019-03-0…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[7]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    96 10.1… hampt… 2011  NA    NA    "@art… https://… https:/… 2019-03-0…
#> # … with 1 more variable: updated_at <chr>
#> 
#> $references[[8]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1   152 10.1… filgu… 2011  NA    NA    @arti… https://… https:/… 2019-03-1…
#> # … with 1 more variable: updated_at <chr>
#> 
#> 
#> attr(,"row.names")
#> [1] 1 2 3 4 5 6 7 8
#> attr(,"class")
#> [1] "mangalSearch"

# Search with custom query (specific column)
search_datasets(query = list(name = "kemp_1977"))

#> Found 1 dataset(s) for query: kemp_1977

#> $id
#> [1] 16
#> 
#> $name
#> [1] "kemp_1977"
#> 
#> $date
#> [1] "1977-01-01T05:00:00.000Z"
#> 
#> $description
#> [1] "Food web for the Crystal River estuary"
#> 
#> $public
#> [1] TRUE
#> 
#> $created_at
#> [1] "2019-02-23T03:02:31.697Z"
#> 
#> $updated_at
#> [1] "2019-02-23T03:02:31.697Z"
#> 
#> $ref_id
#> [1] 16
#> 
#> $user_id
#> [1] 3
#> 
#> $networks
#> $networks[[1]]
#> # A tibble: 1 x 12
#>      id name  date  description public all_interactions created_at updated_at
#>   <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#> 1    80 kemp… 1977… Food web f… TRUE   FALSE            2019-02-2… 2019-02-2…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
#> 
#> 
#> $references
#> $references[[1]]
#> # A tibble: 1 x 11
#>      id doi   author year  jstor pmid  bibtex paper_url data_url created_at
#>   <int> <chr> <chr>  <chr> <chr> <chr> <chr>  <chr>     <chr>    <chr>     
#> 1    16 NA    w. m.… 1977  NA    NA    @inbo… NA        https:/… 2019-02-2…
#> # … with 1 more variable: updated_at <chr>
#> 
#> 
#> attr(,"row.names")
#> [1] 1
#> attr(,"class")
#> [1] "mangalSearch"