Search mangal networks

Search mangal networks

search_networks(query = NULL, verbose = TRUE, ...)

Arguments

query
character keyword used to search (case sensitive) or a sf object used to search in a specific geographical area.

verbose
a logical. Should extra information be reported on progress?

...
arguments from rmangal::get_gen(), ignored for spatial query.

Value

data.frame object with all networks informations Class returned mgSearchNetworks

Examples

search_networks(query="insect%")

#> Found 14 networks

#> # A tibble: 14 x 12
#>       id name  date  description public all_interactions created_at updated_at
#>    <int> <chr> <chr> <chr>       <lgl>  <lgl>            <chr>      <chr>     
#>  1    18 mosq… 1965… Occurence … TRUE   FALSE            2019-02-2… 2019-02-2…
#>  2   909 elbe… 1994… Flower-vis… TRUE   FALSE            2019-02-2… 2019-02-2…
#>  3   948 kato… 1991… Flower and… TRUE   FALSE            2019-02-2… 2019-02-2…
#>  4  1460 corn… 1968… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#>  5  1461 corn… 1968… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#>  6  1471 jiro… 1977… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#>  7  1493 robi… 1948… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#>  8  1501 seif… 1970… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#>  9  1500 seif… 1970… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 10  1502 seif… 1976… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 11  1507 vali… 1966… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 12  1508 vali… 1966… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 13  1510 whit… 1981… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 14  1566 rich… 1922… Food web s… TRUE   FALSE            2019-03-1… 2019-03-1…
#> # … with 4 more variables: dataset_id <int>, user_id <int>, geom.type <chr>,
#> #   geom.coordinates <list>
# NOT RUN {
# Spatial query
library(USAboundaries)
area <- us_states(state="california")
networks_in_area <- search_networks(area)
plot(networks_in_area)
# }