Search for a specific reference with DOI

Search for a specific reference with DOI

search_reference(doi = NULL)

Arguments

doi
character Digital Object Identifier of the article which the network is attached.

Value

data.frame object with the reference corresponding to the query. All datasets and networks related to the publication are also attached to the data.frame. Class returned mgSearchReference

Examples

search_reference(doi = "10.2307/3225248")

#> Found dataset: 
#>  @article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}

#> $id
#> [1] 132
#> 
#> $doi
#> [1] "10.2307/3225248"
#> 
#> $author
#> [1] "koslucher"
#> 
#> $year
#> [1] "1973"
#> 
#> $jstor
#> [1] "https://www.jstor.org/stable/3225248"
#> 
#> $pmid
#> [1] NA
#> 
#> $bibtex
#> [1] "@article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}"
#> 
#> $paper_url
#> [1] "https://www.jstor.org/stable/pdf/3225248.pdf?refreqid=excelsior%3Ae2741b937f74579b7875d7ade8f58f47"
#> 
#> $data_url
#> [1] "https://globalwebdb.com/"
#> 
#> $created_at
#> [1] "2019-03-08T19:27:48.551Z"
#> 
#> $updated_at
#> [1] "2019-03-08T19:27:48.551Z"
#> 
#> $datasets
#> # A tibble: 1 x 9
#>      id name   date   description    public created_at updated_at ref_id user_id
#>   <int> <chr>  <chr>  <chr>          <lgl>  <chr>      <chr>       <int>   <int>
#> 1   133 koslu… 1969-… Food web stru… TRUE   2019-03-0… 2019-03-0…    132       4
#> 
#> $networks
#> # A tibble: 1 x 11
#>      id name  date  geom  description public all_interactions created_at
#>   <int> <chr> <chr> <lgl> <chr>       <lgl>  <lgl>            <chr>     
#> 1  1517 kosl… 1969… NA    Food web s… TRUE   FALSE            2019-03-0…
#> # … with 3 more variables: updated_at <chr>, dataset_id <int>, user_id <int>
#> 
#> attr(,"row.names")
#> [1] 1
#> attr(,"class")
#> [1] "mgSearchReference"