Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

search_taxa(query = NULL, tsn = NULL, gbif = NULL, eol = NULL,
  col = NULL, bold = NULL, ncbi = NULL, original = FALSE,
  verbose = TRUE, ...)

Arguments

query
character query on taxa names

tsn
a numeric. Unique taxonomic identifier from Integrated Taxonomic Information Sytem (https://www.itis.gov/)

gbif
a numeric. Unique taxonomic identifier from Global Biodiversity Information Facility (https://www.gbif.org/)

eol
a numeric. Unique taxonomic identifier from Encyclopedia of Life (https://eol.org/)

col
a numeric. Unique taxonomic identifier from Catalogue of Life (https://www.catalogueoflife.org/)

bold
a numeric. Unique taxonomic identifier from Barcode of Life (http://www.boldsystems.org/)

ncbi
a numeric. Unique taxonomic identifier from National Center for Biotechnology Information (https://www.ncbi.nlm.nih.gov/)

original
a logical. Should query taxa names from the original publication? (see details section, default is set to FALSE, only work for full text search)

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to rmangal::get_gen().

Value

An object of class mgSearchTaxa, which is a data.frame object with all taxa matching the query. All networks in which taxa are involved are also attached to the data.frame.

Details

By default, you are querying taxa based on a homogenized taxonomy (corrected names from TNRS and/or GNR. These taxa might not be the taxon name used in the original publication. In order to query, the original name, set argument original to TRUE.

Examples

search_taxa("Acer")

#> Full text search

#> Found 5 taxa involved in 5 network(s)

#>     id     original_name node_level network_id taxonomy_id
#> 1 2629      Acer negundo      taxon         19           2
#> 2 2630  Acer saccharinum      taxon         19           3
#> 3 8895 Acer shirasawanum      taxon        948        3117
#> 4 8923 Acer ukurunduense      taxon        948        3144
#> 5 8930    Acer japonicum      taxon        948        3151
#>                 created_at               updated_at taxonomy.id
#> 1 2019-02-22T18:48:49.433Z 2019-02-22T18:48:49.433Z           2
#> 2 2019-02-22T18:48:49.465Z 2019-02-22T18:48:49.465Z           3
#> 3 2019-02-25T20:52:53.654Z 2019-02-25T20:52:53.654Z        3117
#> 4 2019-02-25T20:52:54.835Z 2019-02-25T20:52:54.835Z        3144
#> 5 2019-02-25T20:52:55.175Z 2019-02-25T20:52:55.175Z        3151
#>       taxonomy.name taxonomy.ncbi taxonomy.tsn taxonomy.eol taxonomy.bold
#> 1      Acer negundo          4023        28749       583069        100987
#> 2  Acer saccharinum         75745        28757       583072        101028
#> 3 Acer shirasawanum         66216           NA      5613162        485602
#> 4     Acer caudatum        290844           NA      2888949        102168
#> 5    Acer japonicum         47966       837855      2888970        448667
#>   taxonomy.gbif                     taxonomy.col taxonomy.rank
#> 1       3189866 90203e29e2f59e5754167f89b9eba3cc       species
#> 2       3189837 1582ed5db846e241f3e18da418a2a477       species
#> 3       7263086 25b4371fbbdc5b1cd43fad895050cc05       species
#> 4       7263099 1211e426bdec174da9a7526348c26fb9       species
#> 5       8010851 3388e995334d5638aeb11ef84f98b319       species
#>        taxonomy.created_at      taxonomy.updated_at networks.id
#> 1 2019-02-21T21:17:12.585Z 2019-06-14T15:20:36.273Z          19
#> 2 2019-02-21T21:17:12.637Z 2019-06-14T15:20:36.328Z          19
#> 3 2019-02-22T00:31:49.729Z 2019-06-14T15:24:56.217Z         948
#> 4 2019-02-22T00:31:52.729Z 2019-06-14T15:24:57.854Z         948
#> 5 2019-02-22T00:31:53.328Z 2019-06-14T15:24:58.251Z         948
#>               networks.name            networks.date
#> 1 roberson_1929_18990701_19 1899-07-01T04:00:00.000Z
#> 2 roberson_1929_18990701_19 1899-07-01T04:00:00.000Z
#> 3    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#> 4    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#> 5    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#>                                                                                                                                    networks.description
#> 1                                                                         Insects observed pollinating flowers, ten miles of Carlinville, Illinois, USA
#> 2                                                                         Insects observed pollinating flowers, ten miles of Carlinville, Illinois, USA
#> 3 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#> 4 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#> 5 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#>   networks.public networks.all_interactions      networks.created_at
#> 1            TRUE                     FALSE 2019-02-22T18:47:30.657Z
#> 2            TRUE                     FALSE 2019-02-22T18:47:30.657Z
#> 3            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#> 4            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#> 5            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#>        networks.updated_at networks.dataset_id networks.user_id
#> 1 2019-02-22T18:47:30.657Z                   1                3
#> 2 2019-02-22T18:47:30.657Z                   1                3
#> 3 2019-02-25T20:52:09.499Z                  66                2
#> 4 2019-02-25T20:52:09.499Z                  66                2
#> 5 2019-02-25T20:52:09.499Z                  66                2
#>              networks.geom
#> 1 POINT (-89.8818 39.2798)
#> 2 POINT (-89.8818 39.2798)
#> 3 POINT (138.3833 35.5833)
#> 4 POINT (138.3833 35.5833)
#> 5 POINT (138.3833 35.5833)

# Retrieve higher classification
tsn_acer <- search_taxa("Acer")$tsn

#> Full text search

#> Found 5 taxa involved in 5 network(s)

taxize::classification(tsn_acer, db = "itis")

#> Error in UseMethod("as.tsn"): pas de méthode pour 'as.tsn' applicable pour un objet de classe "NULL"