Combine Mangal networks

Combine mgNetworksCollection and mgNetwork objects into a mgNetworksCollection object.

Usage

combine_mgNetworks(...)

Arguments

...

objects of class mgNetworksCollection or mgNetwork or a list #' of objects of these classes.

Value

An object of class mgNetworksCollection.

Examples

# \donttest{
 mg_random_1071 <- get_collection(c(1071))
#> Processing id: 1071 	1/1  
#>                                                                              
#> Data retrieval 100%   
#>                                                                              
#> Data retrieval 100%   
#>                                                                              
#> Processing id: 75 	1/1  
#>                                                                              
#> Processing id: 74 	1/1  
#>                                                                              
 mg_random_1074 <- get_collection(c(1074))
#> Processing id: 1074 	1/1  
#>                                                                              
#> Data retrieval 100%   
#>                                                                              
#> Data retrieval 100%   
#>                                                                              
#> Processing id: 75 	1/1  
#>                                                                              
#> Processing id: 74 	1/1  
#>                                                                              
 combine_mgNetworks(mg_random_1071, mg_random_1074)
#> A collection of 2 networks
#> 
#> * Network #1071 included in dataset #75
#> * Description: structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Bajo
#> * Includes 4 edges and 6 nodes 
#> * Current taxonomic IDs coverage for nodes of this network: 
#>   --> ITIS: 100%, BOLD: 83%, EOL: 100%, COL: 100%, GBIF: 100%, NCBI: 100%
#> * Published in ref # DOI:10.1007/s10641-010-9606-0
#> 
#> * Network #1074 included in dataset #75
#> * Description: structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Fukui
#> * Includes 6 edges and 7 nodes 
#> * Current taxonomic IDs coverage for nodes of this network: 
#>   --> ITIS: 100%, BOLD: 86%, EOL: 100%, COL: 100%, GBIF: 100%, NCBI: 100%
#> * Published in ref # DOI:10.1007/s10641-010-9606-0
#> 
# }