Retrieve all references pertaining to the networks collection or individual network

Retrieve all references pertaining to the networks collection or individual network

get_citation(x)

# S3 method for mgNetwork
get_citation(x)

# S3 method for mgNetworksCollection
get_citation(x)

Arguments

x
an object of class mgNetworksCollection or mgNetworks.

Value

Bibtex entries as a character vector.

Methods (by class)

mgNetwork: Get BibTeX entries for the publication associated to the network.

mgNetworksCollection: Get BibTeX entries for the publication associated to the networks.

Examples

 # network collection
 lagoon_net_collection <- get_collection(search_datasets("lagoon"))

#> Data retrieval 100%   

#>                                                                              

#> Found 2 datasets

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 22 	1/1  

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 53 	1/1  

#>                                                                              

#> Processing id: 86 	1/1  

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 50%   

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 22 	1/1  

#>                                                                              

#> Processing id: 22 	1/1  

#>                                                                              

#> Processing id: 927 	1/1  

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 52 	1/1  

#>                                                                              

#> Processing id: 53 	1/1  

#>                                                                              

#> Processing id: 926 	1/1  

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 52 	1/1  

#>                                                                              

#> Processing id: 53 	1/1  

#>                                                                              

 get_citation(lagoon_net_collection)

#> [1] "@article{Zetina_Rej_n_2003, doi = {10.1016/s0272-7714(02)00410-9}, url = {https://doi.org/10.1016%2Fs0272-7714%2802%2900410-9}, year = 2003, month = {aug}, publisher = {Elsevier {BV}}, volume = {57}, number = {5-6}, pages = {803--815}, author = {Manuel J. Zetina-Rejón and Francisco Arreguí-Sánchez and Ernesto A. Chávez}, title = {Trophic structure and flows of energy in the Huizache{\textendash}Caimanero lagoon complex on the Pacific coast of Mexico},journal = {Estuarine, Coastal and Shelf Science}}"
#> [2] "@book{yanez_1978, Author = {Yáñez-Arancibia, Alejandro}, Editor = {Universidad Nacional Autónoma de México, Centro de Ciencias del Mar y Limnología. Ciudad Universitaria, México, D.F. -- 1a ed.},Title = {Taxonomía, ecología y estructura de las comunidades de peces en lagunas costeras con bocas efímeras del Pacífico de México}, Year = {1978}}"                                                                                                                                                                 

 # individual network
 mg_18 <- get_network_by_id(18)

#> Processing id: 18 	1/1  

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Processing id: 4 	1/1  

#>                                                                              

#> Processing id: 4 	1/1  

#>                                                                              

 get_citation(mg_18)

#> [1] "@article{article, author = {Mosquin, Theodore and Martin, J. E. H.}, year = {1967}, pages = {201-205}, title = {Observations on the pollination biology of plants on Melville Island, N.W.T., Canada}, journal = {Canadian Field Naturalist}, volume = {81}}"