Retrieve mangal network including the nodes and the edges

Retrieve mangal network including the nodes and the edges

get_network_by_id(id, ...)

Arguments

id
numeric mangal ID network

...
arguments from rmangal::get_singletons()

Value

object mgNetwork:

network: list of all generic informations on the network

nodes: data.frame of all nodes with taxonomic informations

edges: data.frame of all edges (ecological interactions), with the attribute used to describe the interaction

dataset: list of all informations on the dataset associated to the network

reference: list of all informations on the original publication

Examples

get_network_by_id(id = 18)

#> $network
#> $network$id
#> [1] 18
#> 
#> $network$name
#> [1] "mosquin_martin_1967_19650731_18"
#> 
#> $network$date
#> [1] "1965-07-31T04:00:00.000Z"
#> 
#> $network$geom
#> $network$geom$type
#> [1] "Point"
#> 
#> $network$geom$coordinates
#> $network$geom$coordinates[[1]]
#> [1] -114.9667
#> 
#> $network$geom$coordinates[[2]]
#> [1] 75
#> 
#> 
#> 
#> $network$description
#> [1] "Occurence of flower-visiting insect on plant species, two miles north of Bailey Point, Melville Island, N.W.T., Canada"
#> 
#> $network$public
#> [1] TRUE
#> 
#> $network$all_interactions
#> [1] FALSE
#> 
#> $network$created_at
#> [1] "2019-02-22T18:38:37.491Z"
#> 
#> $network$updated_at
#> [1] "2019-02-22T18:38:37.491Z"
#> 
#> $network$dataset_id
#> [1] 4
#> 
#> $network$user_id
#> [1] 3
#> 
#> 
#> $nodes
#> # A tibble: 29 x 16
#>       id original_name node_level network_id taxonomy_id created_at updated_at
#>    <int> <chr>         <chr>           <int>       <int> <chr>      <chr>     
#>  1  2599 Spilogona sa… taxon              18        1591 2019-02-2… 2019-02-2…
#>  2  2600 Bombus polar… taxon              18        1956 2019-02-2… 2019-02-2…
#>  3  2601 "Spilogona a… taxon              18        1957 2019-02-2… 2019-02-2…
#>  4  2602 Colias hecla  taxon              18        1540 2019-02-2… 2019-02-2…
#>  5  2603 "Eupogonomyi… taxon              18        1958 2019-02-2… 2019-02-2…
#>  6  2604 "Boloria imp… taxon              18        1959 2019-02-2… 2019-02-2…
#>  7  2605 "Pogonomyiod… taxon              18        1960 2019-02-2… 2019-02-2…
#>  8  2606 Spilogona la… taxon              18        1589 2019-02-2… 2019-02-2…
#>  9  2607 Boreellus at… taxon              18        1596 2019-02-2… 2019-02-2…
#> 10  2608 Spilogona do… taxon              18        1588 2019-02-2… 2019-02-2…
#> # … with 19 more rows, and 9 more variables: taxonomy.id <int>,
#> #   taxonomy.name <chr>, taxonomy.ncbi <int>, taxonomy.tsn <int>,
#> #   taxonomy.eol <int>, taxonomy.bold <int>, taxonomy.status <chr>,
#> #   taxonomy.created_at <chr>, taxonomy.updated_at <chr>
#> 
#> $edges
#> # A tibble: 38 x 20
#>       id node_from node_to date  direction type  method attr_id value geom 
#>    <int>     <int>   <int> <chr> <chr>     <chr> <chr>    <int> <int> <lgl>
#>  1 52630      2599    2617 1965… directed  mutu… field…       3    19 NA   
#>  2 52631      2599    2618 1965… directed  mutu… field…       3    15 NA   
#>  3 52632      2599    2619 1965… directed  mutu… field…       3     1 NA   
#>  4 52633      2599    2620 1965… directed  mutu… field…       3     3 NA   
#>  5 52634      2599    2621 1965… directed  mutu… field…       3     3 NA   
#>  6 52635      2599    2622 1965… directed  mutu… field…       3     2 NA   
#>  7 52636      2600    2623 1965… directed  mutu… field…       3    15 NA   
#>  8 52637      2600    2624 1965… directed  mutu… field…       3     1 NA   
#>  9 52638      2601    2617 1965… directed  mutu… field…       3     5 NA   
#> 10 52639      2601    2618 1965… directed  mutu… field…       3     3 NA   
#> # … with 28 more rows, and 10 more variables: public <lgl>, network_id <int>,
#> #   created_at <chr>, updated_at <chr>, attribute.id <int>,
#> #   attribute.name <chr>, attribute.description <chr>, attribute.unit <chr>,
#> #   attribute.created_at <chr>, attribute.updated_at <chr>
#> 
#> $dataset
#> $dataset$id
#> [1] 4
#> 
#> $dataset$name
#> [1] "mosquin_martin_1967"
#> 
#> $dataset$date
#> [1] "1965-07-31T04:00:00.000Z"
#> 
#> $dataset$description
#> [1] "Occurence of flower-visiting insect on plant species, two miles north of Bailey Point, Melville Island, N.W.T., Canada"
#> 
#> $dataset$public
#> [1] TRUE
#> 
#> $dataset$created_at
#> [1] "2019-02-22T16:45:59.451Z"
#> 
#> $dataset$updated_at
#> [1] "2019-02-22T16:45:59.451Z"
#> 
#> $dataset$ref_id
#> [1] 4
#> 
#> $dataset$user_id
#> [1] 3
#> 
#> 
#> $reference
#> $reference$id
#> [1] 4
#> 
#> $reference$doi
#> NULL
#> 
#> $reference$author
#> [1] "mosquin"
#> 
#> $reference$year
#> [1] "1967"
#> 
#> $reference$jstor
#> NULL
#> 
#> $reference$pmid
#> NULL
#> 
#> $reference$bibtex
#> [1] "@article{article, author = {Mosquin, Theodore and Martin, J. E. H.}, year = {1967}, pages = {201-205}, title = {Observations on the pollination biology of plants on Melville Island, N.W.T., Canada}, journal = {Canadian Field Naturalist}, volume = {81}}"
#> 
#> $reference$paper_url
#> [1] "https://www.biodiversitylibrary.org/pdf4/075601100090098.pdf"
#> 
#> $reference$data_url
#> [1] "https://www.nceas.ucsb.edu/interactionweb/data/plant_pollinator/excel/mosquin_1967.xls"
#> 
#> $reference$created_at
#> [1] "2019-02-22T16:45:54.246Z"
#> 
#> $reference$updated_at
#> [1] "2019-02-22T16:45:54.246Z"
#> 
#> 
#> attr(,"class")
#> [1] "mgNetwork"