Search for specific interactions type (e.g. mutualism)

Search for specific interactions type (e.g. mutualism)

search_interactions(type = avail_type(), expand_node = FALSE,
  verbose = FALSE)

Arguments

type
a character one of the interactions type available in the function avail_type()

expand_node
a logical. Should the function returned more information on nodes involved in the interaction? (default: FALSE, which means that only the mangal ID of nodes are returned)

verbose
a logical. Should extra information be reported on progress?

Value

An object of class mgSearchInteractions, which is a data.frame object with all interaction matching the interaction type provided. All networks in which interactions are involved are also attached to the data.frame.

Examples

search_interactions("competition")

#> # A tibble: 12 x 21
#>        id node_from node_to date  direction type  method attr_id value geom 
#>     <int>     <int>   <int> <chr> <chr>     <chr> <chr>    <int> <int> <lgl>
#>  1  97641     24343   24344 1973… undirect… comp… field…      25     1 NA   
#>  2  97652     24352   24351 1973… undirect… comp… field…      25     1 NA   
#>  3  97709     24370   24378 1973… undirect… comp… field…      25     1 NA   
#>  4  97679     24357   24358 1973… undirect… comp… field…      25     1 NA   
#>  5 100173     25149   25149 1976… undirect… comp… field…      25     1 NA   
#>  6 100174     25149   25150 1976… undirect… comp… field…      25     1 NA   
#>  7 101348     25576   25577 1970… undirect… comp… Field…      25     1 NA   
#>  8 101349     25576   25578 1970… undirect… comp… Field…      25     1 NA   
#>  9 101350     25576   25571 1970… undirect… comp… Field…      25     1 NA   
#> 10 101351     25577   25578 1970… undirect… comp… Field…      25     1 NA   
#> 11 101352     25577   25571 1970… undirect… comp… Field…      25     1 NA   
#> 12 101353     25578   25571 1970… undirect… comp… Field…      25     1 NA   
#> # … with 22 more variables: public <lgl>, network_id <int>, created_at <chr>,
#> #   updated_at <chr>, attribute.id <int>, attribute.name <chr>,
#> #   attribute.description <chr>, attribute.unit <lgl>,
#> #   attribute.created_at <chr>, attribute.updated_at <chr>, networks$id <int>,
#> #   $name <chr>, $date <chr>, $geom.type <chr>, $geom.coordinates <dbl>,
#> #   $description <chr>, $public <lgl>, $all_interactions <lgl>,
#> #   $created_at <chr>, $updated_at <chr>, $dataset_id <int>, $user_id <int>

# Get all networks containing competition
competition_networks <- get_collection(search_interactions("competition"))