Search for specific interactions type (e.g. mutualism)

Search for specific interactions type (e.g. mutualism)

search_interactions(type = avail_type(), expand_node = FALSE,
  verbose = FALSE)

Arguments

type
a character one of the interactions type available in the function avail_type()

expand_node
a logical. Should the function returned more information on nodes involved in the interaction? Default is set to FALSE, which means that only the mangal ID of nodes are returned.

verbose
a logical. Should extra information be reported on progress?

Value

An object of class mgSearchInteractions, which is a data.frame object with all interaction matching the interaction type provided. All networks in which interactions are involved are also attached to the data.frame.

Examples

search_interactions("competition")

#> Simple feature collection with 12 features and 20 fields (with 1 geometry empty)
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: Inf ymin: Inf xmax: -Inf ymax: -Inf
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:
#>        id node_from node_to                     date  direction        type
#> 1   97641     24343   24344 1973-03-01T00:00:00.000Z undirected competition
#> 2   97652     24352   24351 1973-03-01T00:00:00.000Z undirected competition
#> 3   97709     24370   24378 1973-03-01T00:00:00.000Z undirected competition
#> 4   97679     24357   24358 1973-03-01T00:00:00.000Z undirected competition
#> 5  100173     25149   25149 1976-02-01T00:00:00.000Z undirected competition
#> 6  100174     25149   25150 1976-02-01T00:00:00.000Z undirected competition
#> 7  101348     25576   25577 1970-01-01T00:00:00.000Z undirected competition
#> 8  101349     25576   25578 1970-01-01T00:00:00.000Z undirected competition
#> 9  101350     25576   25571 1970-01-01T00:00:00.000Z undirected competition
#> 10 101351     25577   25578 1970-01-01T00:00:00.000Z undirected competition
#>                                       method attr_id value public network_id
#> 1               field observation/experiment      25     1   TRUE       1477
#> 2               field observation/experiment      25     1   TRUE       1478
#> 3               field observation/experiment      25     1   TRUE       1480
#> 4               field observation/experiment      25     1   TRUE       1479
#> 5               field observation/experiment      25     1   TRUE       1502
#> 6               field observation/experiment      25     1   TRUE       1502
#> 7  Field observation, gut content and biblio      25     1   TRUE       1523
#> 8  Field observation, gut content and biblio      25     1   TRUE       1523
#> 9  Field observation, gut content and biblio      25     1   TRUE       1523
#> 10 Field observation, gut content and biblio      25     1   TRUE       1523
#>                  created_at               updated_at attribute.id
#> 1  2019-03-05T15:50:33.839Z 2019-03-05T15:50:33.839Z           25
#> 2  2019-03-05T15:50:38.469Z 2019-03-05T15:50:38.469Z           25
#> 3  2019-03-05T15:50:55.947Z 2019-03-05T15:50:55.947Z           25
#> 4  2019-03-05T15:50:47.266Z 2019-03-05T15:50:47.266Z           25
#> 5  2019-03-06T21:41:55.180Z 2019-03-06T21:41:55.180Z           25
#> 6  2019-03-06T21:41:55.330Z 2019-03-06T21:41:55.330Z           25
#> 7  2019-03-11T20:53:55.935Z 2019-03-11T20:53:55.935Z           25
#> 8  2019-03-11T20:53:56.023Z 2019-03-11T20:53:56.023Z           25
#> 9  2019-03-11T20:53:56.120Z 2019-03-11T20:53:56.120Z           25
#> 10 2019-03-11T20:53:56.250Z 2019-03-11T20:53:56.250Z           25
#>      attribute.name              attribute.description attribute.unit
#> 1  presence/absence Presence or absence of interaction             NA
#> 2  presence/absence Presence or absence of interaction             NA
#> 3  presence/absence Presence or absence of interaction             NA
#> 4  presence/absence Presence or absence of interaction             NA
#> 5  presence/absence Presence or absence of interaction             NA
#> 6  presence/absence Presence or absence of interaction             NA
#> 7  presence/absence Presence or absence of interaction             NA
#> 8  presence/absence Presence or absence of interaction             NA
#> 9  presence/absence Presence or absence of interaction             NA
#> 10 presence/absence Presence or absence of interaction             NA
#>        attribute.created_at     attribute.updated_at        geom networks.id
#> 1  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1477
#> 2  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1478
#> 3  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1480
#> 4  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1479
#> 5  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1502
#> 6  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1502
#> 7  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1523
#> 8  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1523
#> 9  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1523
#> 10 2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY        1523
#>                          networks.name            networks.date
#> 1  menge_sutherland_1976_19730311_1477 1973-03-11T05:00:00.000Z
#> 2  menge_sutherland_1976_19730311_1478 1973-03-11T05:00:00.000Z
#> 3  menge_sutherland_1976_19730311_1480 1973-03-11T05:00:00.000Z
#> 4  menge_sutherland_1976_19730311_1479 1973-03-11T05:00:00.000Z
#> 5           seifert_1979_19760201_1502 1976-02-01T05:00:00.000Z
#> 6           seifert_1979_19760201_1502 1976-02-01T05:00:00.000Z
#> 7                          wilbur_1972 1970-01-01T05:00:00.000Z
#> 8                          wilbur_1972 1970-01-01T05:00:00.000Z
#> 9                          wilbur_1972 1970-01-01T05:00:00.000Z
#> 10                         wilbur_1972 1970-01-01T05:00:00.000Z
#>                                                                                  networks.description
#> 1                 Food web structure of an exposed rocky shore community, Pemaquid point, New England
#> 2               Food web structure of a protected rocky shore community, Grindstone Neck, New England
#> 3                Food web structure of a protected rocky shore community, San Juan Island, Washington
#> 4                       Food web structure of an exposed rocky shore community, Makah Bay, Washington
#> 5                                The insect community of Heliconia bihai in a Venezuelan cloud forest
#> 6                                The insect community of Heliconia bihai in a Venezuelan cloud forest
#> 7  Food web structure of amphibian community within woodland ponds, Edwin S. George Reserve, Michigan
#> 8  Food web structure of amphibian community within woodland ponds, Edwin S. George Reserve, Michigan
#> 9  Food web structure of amphibian community within woodland ponds, Edwin S. George Reserve, Michigan
#> 10 Food web structure of amphibian community within woodland ponds, Edwin S. George Reserve, Michigan
#>    networks.public networks.all_interactions      networks.created_at
#> 1             TRUE                     FALSE 2019-03-05T15:50:31.521Z
#> 2             TRUE                     FALSE 2019-03-05T15:50:34.068Z
#> 3             TRUE                     FALSE 2019-03-05T15:50:47.410Z
#> 4             TRUE                     FALSE 2019-03-05T15:50:38.637Z
#> 5             TRUE                     FALSE 2019-03-06T20:26:41.249Z
#> 6             TRUE                     FALSE 2019-03-06T20:26:41.249Z
#> 7             TRUE                     FALSE 2019-03-11T20:53:23.501Z
#> 8             TRUE                     FALSE 2019-03-11T20:53:23.501Z
#> 9             TRUE                     FALSE 2019-03-11T20:53:23.501Z
#> 10            TRUE                     FALSE 2019-03-11T20:53:23.501Z
#>         networks.updated_at networks.dataset_id networks.user_id
#> 1  2019-03-05T15:50:31.521Z                 105                4
#> 2  2019-03-05T15:50:34.068Z                 105                4
#> 3  2019-03-05T15:50:47.410Z                 105                4
#> 4  2019-03-05T15:50:38.637Z                 105                4
#> 5  2019-03-06T20:26:41.249Z                 119                4
#> 6  2019-03-06T20:26:41.249Z                 119                4
#> 7  2019-03-11T20:53:23.501Z                 139                4
#> 8  2019-03-11T20:53:23.501Z                 139                4
#> 9  2019-03-11T20:53:23.501Z                 139                4
#> 10 2019-03-11T20:53:23.501Z                 139                4
#>                 networks.geom
#> 1    POINT (-69.5098 43.8406)
#> 2  POINT (-68.09012 44.37215)
#> 3    POINT (-123.084 48.5377)
#> 4  POINT (-124.6716 48.32856)
#> 5  POINT (-67.68822 10.34253)
#> 6  POINT (-67.68822 10.34253)
#> 7  POINT (-83.99992 42.45481)
#> 8  POINT (-83.99992 42.45481)
#> 9  POINT (-83.99992 42.45481)
#> 10 POINT (-83.99992 42.45481)

# Get all networks containing competition
competition_networks <- get_collection(search_interactions("competition"))