Search for specific interactions type (e.g. mutualism)

Search for specific interactions type (e.g. mutualism)

search_interactions(type = avail_type(), expand_node = FALSE,
  verbose = FALSE)

Arguments

type
a character one of the interactions type available in the function avail_type()

expand_node
a logical. Should the function returned more information on nodes involved in the interaction? Default is set to FALSE, which means that only the mangal ID of nodes are returned.

verbose
a logical. Should extra information be reported on progress?

Value

An object of class mgSearchInteractions, which is a data.frame object with all interaction matching the interaction type provided. All networks in which interactions are involved are also attached to the data.frame.

Examples

search_interactions("competition")

#> Simple feature collection with 12 features and 20 fields (with 1 geometry empty)
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: Inf ymin: Inf xmax: -Inf ymax: -Inf
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> First 10 features:

#> Warning: corrupt data frame: columns will be truncated or padded with NAs

#>        id node_from node_to                     date  direction        type
#> 1   97641     24343   24344 1973-03-01T00:00:00.000Z undirected competition
#> 2   97652     24352   24351 1973-03-01T00:00:00.000Z undirected competition
#> 3   97709     24370   24378 1973-03-01T00:00:00.000Z undirected competition
#> 4   97679     24357   24358 1973-03-01T00:00:00.000Z undirected competition
#> 5  100173     25149   25149 1976-02-01T00:00:00.000Z undirected competition
#> 6  100174     25149   25150 1976-02-01T00:00:00.000Z undirected competition
#> 7  101348     25576   25577 1970-01-01T00:00:00.000Z undirected competition
#> 8  101349     25576   25578 1970-01-01T00:00:00.000Z undirected competition
#> 9  101350     25576   25571 1970-01-01T00:00:00.000Z undirected competition
#> 10 101351     25577   25578 1970-01-01T00:00:00.000Z undirected competition
#>                                       method attr_id value public network_id
#> 1               field observation/experiment      25     1   TRUE       1477
#> 2               field observation/experiment      25     1   TRUE       1478
#> 3               field observation/experiment      25     1   TRUE       1480
#> 4               field observation/experiment      25     1   TRUE       1479
#> 5               field observation/experiment      25     1   TRUE       1502
#> 6               field observation/experiment      25     1   TRUE       1502
#> 7  Field observation, gut content and biblio      25     1   TRUE       1523
#> 8  Field observation, gut content and biblio      25     1   TRUE       1523
#> 9  Field observation, gut content and biblio      25     1   TRUE       1523
#> 10 Field observation, gut content and biblio      25     1   TRUE       1523
#>                  created_at               updated_at attribute.id
#> 1  2019-03-05T15:50:33.839Z 2019-03-05T15:50:33.839Z           25
#> 2  2019-03-05T15:50:38.469Z 2019-03-05T15:50:38.469Z           25
#> 3  2019-03-05T15:50:55.947Z 2019-03-05T15:50:55.947Z           25
#> 4  2019-03-05T15:50:47.266Z 2019-03-05T15:50:47.266Z           25
#> 5  2019-03-06T21:41:55.180Z 2019-03-06T21:41:55.180Z           25
#> 6  2019-03-06T21:41:55.330Z 2019-03-06T21:41:55.330Z           25
#> 7  2019-03-11T20:53:55.935Z 2019-03-11T20:53:55.935Z           25
#> 8  2019-03-11T20:53:56.023Z 2019-03-11T20:53:56.023Z           25
#> 9  2019-03-11T20:53:56.120Z 2019-03-11T20:53:56.120Z           25
#> 10 2019-03-11T20:53:56.250Z 2019-03-11T20:53:56.250Z           25
#>      attribute.name              attribute.description attribute.unit
#> 1  presence/absence Presence or absence of interaction             NA
#> 2  presence/absence Presence or absence of interaction             NA
#> 3  presence/absence Presence or absence of interaction             NA
#> 4  presence/absence Presence or absence of interaction             NA
#> 5  presence/absence Presence or absence of interaction             NA
#> 6  presence/absence Presence or absence of interaction             NA
#> 7  presence/absence Presence or absence of interaction             NA
#> 8  presence/absence Presence or absence of interaction             NA
#> 9  presence/absence Presence or absence of interaction             NA
#> 10 presence/absence Presence or absence of interaction             NA
#>        attribute.created_at     attribute.updated_at        geom
#> 1  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 2  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 3  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 4  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 5  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 6  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 7  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 8  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 9  2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#> 10 2019-03-01T16:03:38.993Z 2019-03-01T16:03:38.993Z POINT EMPTY
#>                                                                          networks
#> 1                                                             # A tibble: 10 x 11
#> 2        id name  date  description public all_interactions created_at updated_at
#> 3     <int> <fct> <fct> <fct>       <lgl>  <lgl>            <fct>      <fct>     
#> 4   1  1477 meng… 1973… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 5   2  1478 meng… 1973… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 6   3  1480 meng… 1973… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 7   4  1479 meng… 1973… Food web s… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 8   5  1502 seif… 1976… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 9   6  1502 seif… 1976… The insect… TRUE   FALSE            2019-03-0… 2019-03-0…
#> 10  7  1523 wilb… 1970… Food web s… TRUE   FALSE            2019-03-1… 2019-03-1…

# Get all networks containing competition
competition_networks <- get_collection(search_interactions("competition"))