Search for a specific dataset reference with DOI

Search for a specific dataset reference with DOI

search_reference(doi, verbose = TRUE, ...)

Arguments

doi
character a Digital Object Identifier of the article (mandatory arg).

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to rmangal::get_gen().

Value

An object of class mgSearchReference, which is essentially a list that include a wide range of details associated to the reference, including all datasets and networks related to the publication that are included in mangal data base.

Examples

search_reference(doi = "10.2307/3225248")

#> Found dataset: 
#>  @article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}

#> $id
#> [1] 132
#> 
#> $doi
#> [1] 10.2307/3225248
#> Levels: 10.2307/3225248
#> 
#> $author
#> [1] koslucher
#> Levels: koslucher
#> 
#> $year
#> [1] 1973
#> Levels: 1973
#> 
#> $jstor
#> [1] https://www.jstor.org/stable/3225248
#> Levels: https://www.jstor.org/stable/3225248
#> 
#> $pmid
#> [1] NA
#> 
#> $bibtex
#> [1] @article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}
#> Levels: @article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}
#> 
#> $paper_url
#> [1] https://www.jstor.org/stable/pdf/3225248.pdf?refreqid=excelsior%3Ae2741b937f74579b7875d7ade8f58f47
#> Levels: https://www.jstor.org/stable/pdf/3225248.pdf?refreqid=excelsior%3Ae2741b937f74579b7875d7ade8f58f47
#> 
#> $data_url
#> [1] https://globalwebdb.com/
#> Levels: https://globalwebdb.com/
#> 
#> $created_at
#> [1] 2019-03-08T19:27:48.551Z
#> Levels: 2019-03-08T19:27:48.551Z
#> 
#> $updated_at
#> [1] 2019-03-08T19:27:48.551Z
#> Levels: 2019-03-08T19:27:48.551Z
#> 
#> $datasets
#>    id                    name                     date
#> 1 133 koslucher_minshall_1973 1969-11-01T04:00:00.000Z
#>                                                                                   description
#> 1 Food web strucutre of a noterhen cool-desert stream (Deep Creek), Curlew Valley, Idaho-Utah
#>   public               created_at               updated_at ref_id user_id
#> 1   TRUE 2019-03-08T19:27:50.152Z 2019-03-08T19:27:50.152Z    132       4
#> 
#> $networks
#> Simple feature collection with 1 feature and 10 fields (with 1 geometry empty)
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: Inf ymin: Inf xmax: -Inf ymax: -Inf
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>     id                    name                     date
#> 1 1517 koslucher_minshall_1973 1969-11-01T04:00:00.000Z
#>                                                                                   description
#> 1 Food web strucutre of a noterhen cool-desert stream (Deep Creek), Curlew Valley, Idaho-Utah
#>   public all_interactions               created_at               updated_at
#> 1   TRUE            FALSE 2019-03-08T19:27:51.061Z 2019-03-08T19:27:51.061Z
#>   dataset_id user_id        geom
#> 1        133       4 POINT EMPTY
#> 
#> attr(,"row.names")
#> [1] 1
#> attr(,"class")
#> [1] "mgSearchReference"