Query references

Search for a specific reference using a key word or a Digital Object Identifier (DOI). If the query is a character string, then all character columns in the table are searched and the entries for which at least one partial match was found are returned. Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)

search_references(query, doi = NULL, verbose = TRUE, ...)

Arguments

query
either a character string including a single keyword or a list containing a custom query (see details section below). Note that if an empty character string is passed, then all datasets available are returned.

doi
character a Digital Object Identifier (DOI) of the article. Note that query is ignored if doi is specified.

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to httr::GET().

Value

An object of class mgSearchReferences, which is a list that includes a wide range of details associated to the reference, including all datasets and networks related to the publication that are included in Mangal database.

Details

Names of the list should match one of the column names within the table. For the reference table, those are:

id: unique identifier of the reference

first_author: first author

doi: use doi instead

jstor: JSTOR identifier

year: year of publication.

Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.

References

Metadata available at https://mangal-wg.github.io/mangal-api/#references

Examples

search_references(doi = "10.2307/3225248")

#> Data retrieval 100%   

#>                                                                              

#> Found 1 reference(s)

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> $id
#> [1] 132
#> 
#> $doi
#> [1] "10.2307/3225248"
#> 
#> $first_author
#> [1] "koslucher"
#> 
#> $year
#> [1] "1973"
#> 
#> $jstor
#> [1] "https://www.jstor.org/stable/3225248"
#> 
#> $pmid
#> [1] NA
#> 
#> $bibtex
#> [1] "@article{Koslucher_1973, doi = {10.2307/3225248}, url = {https://doi.org/10.2307%2F3225248}, year = 1973, month = {jul}, publisher = {{JSTOR}}, volume = {92}, number = {3}, pages = {441}, author = {Dale G. Koslucher and G. Wayne Minshall}, title = {Food Habits of Some Benthic Invertebrates in a Northern Cool-Desert Stream (Deep Creek, Curlew Valley, Idaho-Utah)}, journal = {Transactions of the American Microscopical Society}}"
#> 
#> $paper_url
#> [1] "https://www.jstor.org/stable/pdf/3225248.pdf?refreqid=excelsior%3Ae2741b937f74579b7875d7ade8f58f47"
#> 
#> $data_url
#> [1] "https://globalwebdb.com/"
#> 
#> $created_at
#> [1] "2019-03-08T19:27:48.551Z"
#> 
#> $updated_at
#> [1] "2019-03-08T19:27:48.551Z"
#> 
#> $datasets
#>    id                    name
#> 1 133 koslucher_minshall_1973
#>                                                                                   description
#> 1 Food web strucutre of a noterhen cool-desert stream (Deep Creek), Curlew Valley, Idaho-Utah
#>   public               created_at               updated_at ref_id user_id
#> 1   TRUE 2019-03-08T19:27:50.152Z 2019-03-08T19:27:50.152Z    132       4
#> 
#> $networks
#> $networks[[1]]
#>     id                    name                     date
#> 1 1517 koslucher_minshall_1973 1969-11-01T00:00:00.000Z
#>                                                                                   description
#> 1 Food web strucutre of a noterhen cool-desert stream (Deep Creek), Curlew Valley, Idaho-Utah
#>   public all_interactions               created_at               updated_at
#> 1   TRUE            FALSE 2019-03-08T19:27:51.061Z 2019-03-08T19:27:51.061Z
#>   dataset_id user_id geom_type geom_lon geom_lat
#> 1        133       4      <NA>       NA       NA
#> 
#> 
#> attr(,"row.names")
#> [1] 1
#> attr(,"class")
#> [1] "mgSearchReferences"

search_references(list(jstor = 3683041))

#> Data retrieval 100%   

#>                                                                              

#> Found 1 reference(s)

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> $id
#> [1] 9
#> 
#> $doi
#> [1] "10.2307/3683041"
#> 
#> $first_author
#> [1] "elberling"
#> 
#> $year
#> [1] "1999"
#> 
#> $jstor
#> [1] "3683041"
#> 
#> $pmid
#> [1] NA
#> 
#> $bibtex
#> [1] "@article{10.2307/3683041, ISSN = {09067590, 16000587}, URL = {http://www.jstor.org/stable/3683041}, author = {Heidi Elberling and Jens M. Olesen}, journal = {Ecography}, number = {3}, pages = {314-323}, publisher = {[Nordic Society Oikos, Wiley]}, title = {The Structure of a High Latitude Plant-Flower Visitor System: The Dominance of Flies}, volume = {22}, year = {1999}}"
#> 
#> $paper_url
#> [1] "https://www.jstor.org/stable/3683041?seq=1#page_scan_tab_contents"
#> 
#> $data_url
#> [1] "https://www.nceas.ucsb.edu/interactionweb/data/plant_pollinator/excel/elberling&olesen_1999.xls"
#> 
#> $created_at
#> [1] "2019-02-22T20:09:13.872Z"
#> 
#> $updated_at
#> [1] "2019-02-22T20:09:13.872Z"
#> 
#> $datasets
#>   id                  name
#> 1  9 elberling_olesen_1999
#>                                                   description public
#> 1 Flower-visiting insect at Mt. Latnjatjarro, northern Sweden   TRUE
#>                 created_at               updated_at ref_id user_id
#> 1 2019-02-22T20:09:17.994Z 2019-02-22T20:09:17.994Z      9       2
#> 
#> $networks
#> $networks[[1]]
#>    id                               name                     date
#> 1 909 elberling_olesen_1999_19940823_909 1994-08-23T00:00:00.000Z
#>                                                   description public
#> 1 Flower-visiting insect at Mt. Latnjatjarro, northern Sweden   TRUE
#>   all_interactions               created_at               updated_at dataset_id
#> 1            FALSE 2019-02-24T22:21:32.444Z 2019-02-24T22:21:32.444Z          9
#>   user_id geom_type geom_lon geom_lat
#> 1       2     Point     18.5    68.35
#> 
#> 
#> attr(,"row.names")
#> [1] 1
#> attr(,"class")
#> [1] "mgSearchReferences"

search_references(list(year = 2010))

#> Data retrieval 100%   

#>                                                                              

#> Found 5 reference(s)

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> Data retrieval 100%   

#>                                                                              

#> $id
#> [1]  74  76 111  85 164
#> 
#> $doi
#> [1] "10.1007/s10641-010-9606-0"        "10.3989/scimar.2011.75n2309"     
#> [3] "10.1071/mf09202"                  NA                                
#> [5] "10.1111/j.1461-0248.2009.01437.x"
#> 
#> $first_author
#> [1] "ricciardi"        "ronaldo angelini" "rayner"           "angelini"        
#> [5] "kaiser-bunbury"  
#> 
#> $year
#> [1] "2010" "2010" "2010" "2010" "2010"
#> 
#> $jstor
#> [1] NA NA NA NA NA
#> 
#> $pmid
#> [1] NA NA NA NA NA
#> 
#> $bibtex
#> [1] "@article{Ricciardi_2010,doi = {10.1007/s10641-010-9606-0},url = {https://doi.org/10.1007%2Fs10641-010-9606-0},year = 2010,month = {feb},publisher = {Springer Nature},volume = {87},number = {4},pages = {333--347},author = {Francesco Ricciardi and Massimo Boyer and Jeff Ollerton},title = {Assemblage and interaction structure of the anemonefish-anemone mutualism across the Manado region of Sulawesi, Indonesia},journal = {Environmental Biology of Fishes}}"                                                                          
#> [2] "@article{Angelini_2010, doi = {10.3989/scimar.2011.75n2309}, url = {https://doi.org/10.3989%2Fscimar.2011.75n2309}, year = 2010, month = {nov}, publisher = {Departmento de Publicaciones del {CSIC}}, volume = {75}, number = {2}, pages = {309--319}, author = {Ronaldo Angelini and Filomena Vaz-Velho}, title = {Ecosystem structure and trophic analysis of Angolan fishery landings}, journal = {Scientia Marina}}"                                                                                                                         
#> [3] "@article{Rayner_2010, doi = {10.1071/mf09202}, url = {https://doi.org/10.1071%2Fmf09202}, year = 2010, publisher = {{CSIRO} Publishing}, volume = {61}, number = {8}, pages = {909}, author = {Thomas S. Rayner and Bradley J. Pusey and Richard G. Pearson and Paul C. Godfrey}, title = {Food web dynamics in an Australian Wet Tropics river}, journal = {Marine and Freshwater Research}}"                                                                                                                                                    
#> [4] "@article{angelini2010mixed, title={Mixed food web control and stability in a Cerrado river (Brazil)}, author={Angelini, RONALDO and Aloisio, GUSTAVO RIBEIRO and Carvalho, ADRIANA ROSA}, journal={Pan-American Journal of Aquatic Sciences}, volume={5}, number={3}, pages={421--431}, year={2010}}"                                                                                                                                                                                                                                             
#> [5] "@article{Kaiser_Bunbury_2010, doi = {10.1111/j.1461-0248.2009.01437.x}, url = {https://doi.org/10.1111%2Fj.1461-0248.2009.01437.x}, year = 2010, month = {apr}, publisher = {Wiley}, volume = {13}, number = {4}, pages = {442--452}, author = {Christopher N. Kaiser-Bunbury and Stefanie Muff and Jane Memmott and Christine B. Müller and Amedeo Caflisch}, title = {The robustness of pollination networks to the loss of species and interactions: a quantitative approach incorporating pollinator behaviour}, journal = {Ecology Letters}}"
#> 
#> $paper_url
#> [1] "https://link.springer.com/article/10.1007%2Fs10641-010-9606-0"            
#> [2] "https://doi.org/10.3989%2Fscimar.2011.75n2309"                            
#> [3] "http://www.publish.csiro.au/mf/MF09202"                                   
#> [4] "https://panamjas.org/pdf_artigos/PANAMJAS_5(3)_421-431.pdf"               
#> [5] "https://onlinelibrary.wiley.com/doi/full/10.1111/j.1461-0248.2009.01437.x"
#> 
#> $data_url
#> [1] "https://www.nceas.ucsb.edu/interactionweb/data/anemone_fish/ricciardi-et-al-2010.xls"
#> [2] "https://globalwebdb.com/"                                                            
#> [3] "https://globalwebdb.com/"                                                            
#> [4] "https://globalwebdb.com/"                                                            
#> [5] "http://www.web-of-life.es/2.0/map.php"                                               
#> 
#> $created_at
#> [1] "2019-02-26T20:11:58.429Z" "2019-02-26T21:18:52.787Z"
#> [3] "2019-03-06T14:49:40.733Z" "2019-02-28T17:26:42.053Z"
#> [5] "2019-03-21T19:21:35.795Z"
#> 
#> $updated_at
#> [1] "2019-02-26T20:11:58.429Z" "2019-02-26T21:18:52.787Z"
#> [3] "2019-03-06T14:49:40.733Z" "2019-02-28T17:26:42.053Z"
#> [5] "2019-03-21T19:21:35.795Z"
#> 
#> $datasets
#>    id                      name
#> 1  75            ricciardi_2010
#> 2  77             angelini_2010
#> 3 112         rayner_et_al_2010
#> 4  86       angelini_et_al_2010
#> 5 167 kaiser-bunbury_et_al_2010
#>                                                                                       description
#> 1 structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia
#> 2                                                        Food web of the Angolan fishery landings
#> 3                            Food web structure of a wet Tropic river (Mulgrave River), Australia
#> 4         Interaction of a mixed food web in the Corrente River in the Paraná River Basin, Brazil
#> 5             Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#>   public               created_at               updated_at ref_id user_id
#> 1   TRUE 2019-02-26T20:12:00.676Z 2019-02-26T20:12:00.676Z     74       2
#> 2   TRUE 2019-02-26T21:18:55.909Z 2019-02-26T21:18:55.909Z     76       3
#> 3   TRUE 2019-03-06T14:49:42.906Z 2019-03-06T14:49:42.906Z    111       4
#> 4   TRUE 2019-02-28T17:26:45.617Z 2019-02-28T17:26:45.617Z     85       4
#> 5   TRUE 2019-03-21T19:21:38.571Z 2019-03-21T19:21:38.571Z    164       4
#> 
#> $networks
#> $networks[[1]]
#>      id                         name                     date
#> 1  1071 ricciardi_2010_20060701_1071 2006-07-01T00:00:00.000Z
#> 2  1070 ricciardi_2010_20060701_1070 2006-07-01T00:00:00.000Z
#> 3  1072 ricciardi_2010_20060701_1072 2006-07-01T00:00:00.000Z
#> 4  1074 ricciardi_2010_20060701_1074 2006-07-01T00:00:00.000Z
#> 5  1079 ricciardi_2010_20060701_1079 2006-07-01T00:00:00.000Z
#> 6  1080 ricciardi_2010_20060701_1080 2006-07-01T00:00:00.000Z
#> 7  1073 ricciardi_2010_20060701_1073 2006-07-01T00:00:00.000Z
#> 8  1078 ricciardi_2010_20060701_1078 2006-07-01T00:00:00.000Z
#> 9  1083 ricciardi_2010_20060701_1083 2006-07-01T00:00:00.000Z
#> 10 1084 ricciardi_2010_20060701_1084 2006-07-01T00:00:00.000Z
#> 11 1081 ricciardi_2010_20060701_1081 2006-07-01T00:00:00.000Z
#> 12 1082 ricciardi_2010_20060701_1082 2006-07-01T00:00:00.000Z
#> 13 1077 ricciardi_2010_20060701_1077 2006-07-01T00:00:00.000Z
#> 14 1085 ricciardi_2010_20060701_1085 2006-07-01T00:00:00.000Z
#> 15 1075 ricciardi_2010_20060701_1075 2006-07-01T00:00:00.000Z
#> 16 1076 ricciardi_2010_20060701_1076 2006-07-01T00:00:00.000Z
#>                                                                                                     description
#> 1        structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Bajo
#> 2      structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Bahowo
#> 3  structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Batu Kapal
#> 4       structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Fukui
#> 5    structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Mandolin
#> 6       structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Mapia
#> 7       structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Bualo
#> 8    structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Likuan 1
#> 9     structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  T. Kopi
#> 10   structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  T. Pisok
#> 11       structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Mike
#> 12    structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Sachiko
#> 13  structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Kasegaran
#> 14      structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Timur
#> 15   structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Jalan m.
#> 16    structure of local anemonefish-anemone networks across the Manado region of Sulawesi, Indonesia,  Johnson
#>    public all_interactions               created_at               updated_at
#> 1    TRUE            FALSE 2019-02-26T20:12:57.710Z 2019-02-26T20:12:57.710Z
#> 2    TRUE            FALSE 2019-02-26T20:12:52.618Z 2019-02-26T20:12:52.618Z
#> 3    TRUE            FALSE 2019-02-26T20:13:00.989Z 2019-02-26T20:13:00.989Z
#> 4    TRUE            FALSE 2019-02-26T20:13:06.990Z 2019-02-26T20:13:06.990Z
#> 5    TRUE            FALSE 2019-02-26T20:13:22.816Z 2019-02-26T20:13:22.816Z
#> 6    TRUE            FALSE 2019-02-26T20:13:25.096Z 2019-02-26T20:13:25.096Z
#> 7    TRUE            FALSE 2019-02-26T20:13:02.567Z 2019-02-26T20:13:02.567Z
#> 8    TRUE            FALSE 2019-02-26T20:13:18.623Z 2019-02-26T20:13:18.623Z
#> 9    TRUE            FALSE 2019-02-26T20:13:36.429Z 2019-02-26T20:13:36.429Z
#> 10   TRUE            FALSE 2019-02-26T20:13:39.735Z 2019-02-26T20:13:39.735Z
#> 11   TRUE            FALSE 2019-02-26T20:13:28.721Z 2019-02-26T20:13:28.721Z
#> 12   TRUE            FALSE 2019-02-26T20:13:31.516Z 2019-02-26T20:13:31.516Z
#> 13   TRUE            FALSE 2019-02-26T20:13:15.768Z 2019-02-26T20:13:15.768Z
#> 14   TRUE            FALSE 2019-02-26T20:13:44.536Z 2019-02-26T20:13:44.536Z
#> 15   TRUE            FALSE 2019-02-26T20:13:09.431Z 2019-02-26T20:13:09.431Z
#> 16   TRUE            FALSE 2019-02-26T20:13:13.923Z 2019-02-26T20:13:13.923Z
#>    dataset_id user_id geom_type geom_lon geom_lat
#> 1          75       2     Point  124.778    1.787
#> 2          75       2     Point  124.819    1.585
#> 3          75       2     Point  124.704    1.616
#> 4          75       2     Point  124.793    1.767
#> 5          75       2     Point  124.735    1.608
#> 6          75       2     Point  124.744    1.455
#> 7          75       2     Point  124.752    1.606
#> 8          75       2     Point  124.773    1.594
#> 9          75       2     Point  124.698    1.651
#> 10         75       2     Point  124.792    1.578
#> 11         75       2     Point  124.743    1.639
#> 12         75       2     Point  124.768     1.63
#> 13         75       2     Point  124.793    1.464
#> 14         75       2     Point  124.786    1.612
#> 15         75       2     Point  124.756    1.616
#> 16         75       2     Point  124.864    1.608
#> 
#> $networks[[2]]
#>     id                        name                     date
#> 1 1101 angelini_2010_20100601_1101 2010-06-01T00:00:00.000Z
#>                                description public all_interactions
#> 1 Food web of the Angolan fishery landings   TRUE            FALSE
#>                 created_at               updated_at dataset_id user_id
#> 1 2019-02-26T21:19:01.124Z 2019-02-26T21:19:01.124Z         77       3
#>   geom_type
#> 1   Polygon
#>                                                                                   geom_lon
#> 1 10.95337, 11.62354, 11.64551, 12.12891, 13.60107, 11.95312, 10.70068, 10.95337, 10.95337
#>                                                                                          geom_lat
#> 1 -17.25000, -17.25000, -15.53838, -13.96605, -11.13529, -5.00000, -5.00000, -17.25000, -17.25000
#> 
#> $networks[[3]]
#>     id                            name                     date
#> 1 1488 rayner_et_al_2010_20040401_1488 2004-04-01T00:00:00.000Z
#>                                                            description public
#> 1 Food web structure of a wet Tropic river (Mulgrave River), Australia   TRUE
#>   all_interactions               created_at               updated_at dataset_id
#> 1            FALSE 2019-03-06T14:49:45.109Z 2019-03-06T14:49:45.109Z        112
#>   user_id geom_type geom_lon  geom_lat
#> 1       4     Point 145.8646 -17.12026
#> 
#> $networks[[4]]
#>     id                              name                     date
#> 1 1448 angelini_et_al_2010_20030501_1448 2003-05-01T00:00:00.000Z
#>                                                                       description
#> 1 Interaction of a mixed food web in the Corrente River in the Paraná River Basin
#>   public all_interactions               created_at               updated_at
#> 1   TRUE            FALSE 2019-02-28T17:26:47.593Z 2019-02-28T17:26:47.593Z
#>   dataset_id user_id geom_type geom_lon  geom_lat
#> 1         86       4     Point -52.0155 -18.56476
#> 
#> $networks[[5]]
#>      id                         name                     date
#> 1  1599  kaiser-bunbury_et_al_2010_1 2003-09-01T00:00:00.000Z
#> 2  1600  kaiser-bunbury_et_al_2010_2 2003-09-01T00:00:00.000Z
#> 3  1601  kaiser-bunbury_et_al_2010_3 2003-09-01T00:00:00.000Z
#> 4  1602  kaiser-bunbury_et_al_2010_4 2003-09-01T00:00:00.000Z
#> 5  1603  kaiser-bunbury_et_al_2010_5 2003-09-01T00:00:00.000Z
#> 6  1604  kaiser-bunbury_et_al_2010_6 2003-09-01T00:00:00.000Z
#> 7  1605  kaiser-bunbury_et_al_2010_7 2003-09-01T00:00:00.000Z
#> 8  1606  kaiser-bunbury_et_al_2010_8 2003-09-01T00:00:00.000Z
#> 9  1607  kaiser-bunbury_et_al_2010_9 2003-09-01T00:00:00.000Z
#> 10 1608 kaiser-bunbury_et_al_2010_10 2003-09-01T00:00:00.000Z
#> 11 1609 kaiser-bunbury_et_al_2010_11 2003-09-01T00:00:00.000Z
#> 12 1610 kaiser-bunbury_et_al_2010_12 2003-09-01T00:00:00.000Z
#> 13 1611 kaiser-bunbury_et_al_2010_13 2003-09-01T00:00:00.000Z
#> 14 1612 kaiser-bunbury_et_al_2010_14 2003-09-01T00:00:00.000Z
#> 15 1613 kaiser-bunbury_et_al_2010_15 2003-09-01T00:00:00.000Z
#> 16 1614 kaiser-bunbury_et_al_2010_16 2003-09-01T00:00:00.000Z
#> 17 1615 kaiser-bunbury_et_al_2010_17 2003-09-01T00:00:00.000Z
#> 18 1616 kaiser-bunbury_et_al_2010_18 2003-09-01T00:00:00.000Z
#> 19 1617 kaiser-bunbury_et_al_2010_19 2003-09-01T00:00:00.000Z
#> 20 1618 kaiser-bunbury_et_al_2010_20 2003-09-01T00:00:00.000Z
#> 21 1619 kaiser-bunbury_et_al_2010_21 2003-09-01T00:00:00.000Z
#> 22 1620 kaiser-bunbury_et_al_2010_22 2003-09-01T00:00:00.000Z
#> 23 1621 kaiser-bunbury_et_al_2010_23 2003-09-01T00:00:00.000Z
#> 24 1622 kaiser-bunbury_et_al_2010_24 2003-09-01T00:00:00.000Z
#>                                                                            description
#> 1  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 2  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 3  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 4  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 5  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 6  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 7  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 8  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 9  Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 10 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 11 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 12 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 13 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 14 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 15 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 16 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 17 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 18 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 19 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 20 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 21 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 22 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 23 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#> 24 Plant-pollinator system at two sites in Black River Gorges National Park, Mauritius
#>    public all_interactions               created_at               updated_at
#> 1    TRUE            FALSE 2019-03-21T19:23:49.316Z 2019-03-21T19:23:49.316Z
#> 2    TRUE            FALSE 2019-03-21T19:24:08.441Z 2019-03-21T19:24:08.441Z
#> 3    TRUE            FALSE 2019-03-21T19:24:29.240Z 2019-03-21T19:24:29.240Z
#> 4    TRUE            FALSE 2019-03-21T19:24:56.896Z 2019-03-21T19:24:56.896Z
#> 5    TRUE            FALSE 2019-03-21T19:25:27.173Z 2019-03-21T19:25:27.173Z
#> 6    TRUE            FALSE 2019-03-21T19:26:00.215Z 2019-03-21T19:26:00.215Z
#> 7    TRUE            FALSE 2019-03-21T19:26:20.237Z 2019-03-21T19:26:20.237Z
#> 8    TRUE            FALSE 2019-03-21T19:26:47.151Z 2019-03-21T19:26:47.151Z
#> 9    TRUE            FALSE 2019-03-21T19:27:05.322Z 2019-03-21T19:27:05.322Z
#> 10   TRUE            FALSE 2019-03-21T19:27:21.613Z 2019-03-21T19:27:21.613Z
#> 11   TRUE            FALSE 2019-03-21T19:27:32.137Z 2019-03-21T19:27:32.137Z
#> 12   TRUE            FALSE 2019-03-21T19:27:41.446Z 2019-03-21T19:27:41.446Z
#> 13   TRUE            FALSE 2019-03-21T19:27:53.533Z 2019-03-21T19:27:53.533Z
#> 14   TRUE            FALSE 2019-03-21T19:28:03.939Z 2019-03-21T19:28:03.939Z
#> 15   TRUE            FALSE 2019-03-21T19:28:22.521Z 2019-03-21T19:28:22.521Z
#> 16   TRUE            FALSE 2019-03-21T19:28:47.688Z 2019-03-21T19:28:47.688Z
#> 17   TRUE            FALSE 2019-03-21T19:29:08.673Z 2019-03-21T19:29:08.673Z
#> 18   TRUE            FALSE 2019-03-21T19:29:27.829Z 2019-03-21T19:29:27.829Z
#> 19   TRUE            FALSE 2019-03-21T19:29:42.942Z 2019-03-21T19:29:42.942Z
#> 20   TRUE            FALSE 2019-03-21T19:29:51.654Z 2019-03-21T19:29:51.654Z
#> 21   TRUE            FALSE 2019-03-21T19:29:59.208Z 2019-03-21T19:29:59.208Z
#> 22   TRUE            FALSE 2019-03-21T19:30:06.009Z 2019-03-21T19:30:06.009Z
#> 23   TRUE            FALSE 2019-03-21T19:30:18.766Z 2019-03-21T19:30:18.766Z
#> 24   TRUE            FALSE 2019-03-21T19:30:30.248Z 2019-03-21T19:30:30.248Z
#>    dataset_id user_id geom_type geom_lon  geom_lat
#> 1         167       4     Point 57.45094 -20.42637
#> 2         167       4     Point 57.45094 -20.42637
#> 3         167       4     Point 57.45094 -20.42637
#> 4         167       4     Point 57.45094 -20.42637
#> 5         167       4     Point 57.45094 -20.42637
#> 6         167       4     Point 57.45094 -20.42637
#> 7         167       4     Point 57.45094 -20.42637
#> 8         167       4     Point 57.45094 -20.42637
#> 9         167       4     Point 57.45094 -20.42637
#> 10        167       4     Point 57.45094 -20.42637
#> 11        167       4     Point 57.45094 -20.42637
#> 12        167       4     Point 57.45094 -20.42637
#> 13        167       4     Point 57.45094 -20.42637
#> 14        167       4     Point 57.45094 -20.42637
#> 15        167       4     Point 57.45094 -20.42637
#> 16        167       4     Point 57.45094 -20.42637
#> 17        167       4     Point 57.45094 -20.42637
#> 18        167       4     Point 57.45094 -20.42637
#> 19        167       4     Point 57.45094 -20.42637
#> 20        167       4     Point 57.45094 -20.42637
#> 21        167       4     Point 57.45094 -20.42637
#> 22        167       4     Point 57.45094 -20.42637
#> 23        167       4     Point 57.45094 -20.42637
#> 24        167       4     Point 57.45094 -20.42637
#> 
#> 
#> attr(,"row.names")
#> [1] 1 2 3 4 5
#> attr(,"class")
#> [1] "mgSearchReferences"