Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

search_taxa(query = NULL, original = FALSE, verbose = TRUE, ...)

Arguments

query
character specific taxon / keyword / or tsn, eol (page ID), bold & ncbi IDs (mandatory arg).

original
a logical. Should query taxa names from the original publication? (see details section, default is set to FALSE)

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to rmangal::get_gen().

Value

An object of class mgSearchTaxa, which is a data.frame object with all taxa matching the query. All networks in which taxa are involved are also attached to the data.frame.

Details

By default, you are querying taxa based on a homogenized taxonomy (corrected names from TNRS and/or GNR. These taxa might not be the taxon name used in the original publication. In order to query, the original name, set argument original to TRUE.

Examples

search_taxa("Acer")

#> Found 5 taxa

#> Warning: corrupt data frame: columns will be truncated or padded with NAs

#>     id              name   ncbi    tsn     eol   bold    gbif
#> 1    2      Acer negundo   4023  28749  583069 100987 3189866
#> 2    3  Acer saccharinum  75745  28757  583072 101028 3189837
#> 3 3117 Acer shirasawanum  66216     NA 5613162 485602 7263086
#> 4 3144     Acer caudatum 290844     NA 2888949 102168 7263099
#> 5 3151    Acer japonicum  47966 837855 2888970 448667 8010851
#>                                col    rank               created_at
#> 1 90203e29e2f59e5754167f89b9eba3cc species 2019-02-21T21:17:12.585Z
#> 2 1582ed5db846e241f3e18da418a2a477 species 2019-02-21T21:17:12.637Z
#> 3 25b4371fbbdc5b1cd43fad895050cc05 species 2019-02-22T00:31:49.729Z
#> 4 1211e426bdec174da9a7526348c26fb9 species 2019-02-22T00:31:52.729Z
#> 5 3388e995334d5638aeb11ef84f98b319 species 2019-02-22T00:31:53.328Z
#>                 updated_at     original_name
#> 1 2019-06-14T15:20:36.273Z      Acer negundo
#> 2 2019-06-14T15:20:36.328Z  Acer saccharinum
#> 3 2019-06-14T15:24:56.217Z Acer shirasawanum
#> 4 2019-06-14T15:24:57.854Z Acer ukurunduense
#> 5 2019-06-14T15:24:58.251Z    Acer japonicum
#>                                                                        networks
#> 1                                                            # A tibble: 5 x 11
#> 2      id name  date  description public all_interactions created_at updated_at
#> 3   <int> <fct> <fct> <fct>       <lgl>  <lgl>            <fct>      <fct>     
#> 4 1    19 robe… 1899… Insects ob… TRUE   FALSE            2019-02-2… 2019-02-2…
#> 5 2    19 robe… 1899… Insects ob… TRUE   FALSE            2019-02-2… 2019-02-2…

# Retrieve higher classification
tsn_acer <- search_taxa("Acer")$tsn

#> Found 5 taxa

taxize::classification(tsn_acer, db = "itis")

#> $`28749`
#>               name          rank     id
#> 1          Plantae       kingdom 202422
#> 2    Viridiplantae    subkingdom 954898
#> 3     Streptophyta  infrakingdom 846494
#> 4      Embryophyta superdivision 954900
#> 5     Tracheophyta      division 846496
#> 6  Spermatophytina   subdivision 846504
#> 7    Magnoliopsida         class  18063
#> 8          Rosanae    superorder 846548
#> 9       Sapindales         order  28643
#> 10     Sapindaceae        family  28657
#> 11            Acer         genus  28727
#> 12    Acer negundo       species  28749
#> 
#> $`28757`
#>                name          rank     id
#> 1           Plantae       kingdom 202422
#> 2     Viridiplantae    subkingdom 954898
#> 3      Streptophyta  infrakingdom 846494
#> 4       Embryophyta superdivision 954900
#> 5      Tracheophyta      division 846496
#> 6   Spermatophytina   subdivision 846504
#> 7     Magnoliopsida         class  18063
#> 8           Rosanae    superorder 846548
#> 9        Sapindales         order  28643
#> 10      Sapindaceae        family  28657
#> 11             Acer         genus  28727
#> 12 Acer saccharinum       species  28757
#> 
#> $<NA>
#> [1] NA
#> 
#> $<NA>
#> [1] NA
#> 
#> $`837855`
#>               name          rank     id
#> 1          Plantae       kingdom 202422
#> 2    Viridiplantae    subkingdom 954898
#> 3     Streptophyta  infrakingdom 846494
#> 4      Embryophyta superdivision 954900
#> 5     Tracheophyta      division 846496
#> 6  Spermatophytina   subdivision 846504
#> 7    Magnoliopsida         class  18063
#> 8          Rosanae    superorder 846548
#> 9       Sapindales         order  28643
#> 10     Sapindaceae        family  28657
#> 11            Acer         genus  28727
#> 12  Acer japonicum       species 837855
#> 
#> attr(,"class")
#> [1] "classification"
#> attr(,"db")
#> [1] "itis"