Search network by taxa using keyword or speicific taxonomic identifiers.

Search network by taxa using keyword or speicific taxonomic identifiers.

search_taxa(query, tsn = NULL, gbif = NULL, eol = NULL, col = NULL,
  bold = NULL, ncbi = NULL, original = FALSE, verbose = TRUE, ...)

Arguments

query
either a character string including a single keyword or a list containing a custom query (see details section below). Note that if an empty character string is passed, then all datasets available are returned.

tsn
a numeric. Unique taxonomic identifier from Integrated Taxonomic Information System (https://www.itis.gov/).

gbif
a numeric. Unique taxonomic identifier from Global Biodiversity Information Facility (https://www.gbif.org/).

eol
a numeric. Unique taxonomic identifier from Encyclopedia of Life (https://eol.org/).

col
a numeric. Unique taxonomic identifier from Catalogue of Life (https://www.catalogueoflife.org/).

bold
a numeric. Unique taxonomic identifier from Barcode of Life (http://www.boldsystems.org/).

ncbi
a numeric. Unique taxonomic identifier from National Center for Biotechnology Information (https://www.ncbi.nlm.nih.gov/).

original
a logical. Should query taxa names from the original publication? (see details section, default is set to FALSE, only work for full text search)

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to httr::GET()

Value

An object of class mgSearchTaxa, which is a data.frame object with all taxa matching the query. All networks in which taxa are involved are also attached to the data.frame.

Details

By default, the user query the taxonomy table and the based a homogenized taxonomy (valid names from TNRS and/or GNR. These taxa might not be the taxon name documented in the original publication. In order to query, the original name, set argument original to TRUE.

Note that the validation of taxon names was performed by an automated procedure and if there is any doubt, the original names recorded by authors should be regarded as the most reliable information. Please report any issue related to taxonomy at https://github.com/mangal-wg/mangal-datasets/issues.

References

https://mangal-wg.github.io/mangal-api/#nodes https://mangal-wg.github.io/mangal-api/#taxonomy

Examples

search_taxa("Acer")

#>     id     original_name node_level network_id taxonomy_id
#> 1 2629      Acer negundo      taxon         19           2
#> 2 2630  Acer saccharinum      taxon         19           3
#> 3 8895 Acer shirasawanum      taxon        948        3117
#> 4 8923 Acer ukurunduense      taxon        948        3144
#> 5 8930    Acer japonicum      taxon        948        3151
#>                 created_at               updated_at taxonomy.id
#> 1 2019-02-22T18:48:49.433Z 2019-02-22T18:48:49.433Z           2
#> 2 2019-02-22T18:48:49.465Z 2019-02-22T18:48:49.465Z           3
#> 3 2019-02-25T20:52:53.654Z 2019-02-25T20:52:53.654Z        3117
#> 4 2019-02-25T20:52:54.835Z 2019-02-25T20:52:54.835Z        3144
#> 5 2019-02-25T20:52:55.175Z 2019-02-25T20:52:55.175Z        3151
#>       taxonomy.name taxonomy.ncbi taxonomy.tsn taxonomy.eol taxonomy.bold
#> 1      Acer negundo          4023        28749       583069        100987
#> 2  Acer saccharinum         75745        28757       583072        101028
#> 3 Acer shirasawanum         66216           NA      5613162        485602
#> 4     Acer caudatum        290844           NA      2888949        102168
#> 5    Acer japonicum         47966       837855      2888970        448667
#>   taxonomy.gbif                     taxonomy.col taxonomy.rank
#> 1       3189866 90203e29e2f59e5754167f89b9eba3cc       species
#> 2       3189837 1582ed5db846e241f3e18da418a2a477       species
#> 3       7263086 25b4371fbbdc5b1cd43fad895050cc05       species
#> 4       7263099 1211e426bdec174da9a7526348c26fb9       species
#> 5       8010851 3388e995334d5638aeb11ef84f98b319       species
#>        taxonomy.created_at      taxonomy.updated_at networks.id
#> 1 2019-02-21T21:17:12.585Z 2019-06-14T15:20:36.273Z          19
#> 2 2019-02-21T21:17:12.637Z 2019-06-14T15:20:36.328Z          19
#> 3 2019-02-22T00:31:49.729Z 2019-06-14T15:24:56.217Z         948
#> 4 2019-02-22T00:31:52.729Z 2019-06-14T15:24:57.854Z         948
#> 5 2019-02-22T00:31:53.328Z 2019-06-14T15:24:58.251Z         948
#>               networks.name            networks.date
#> 1 roberson_1929_18990701_19 1899-07-01T04:00:00.000Z
#> 2 roberson_1929_18990701_19 1899-07-01T04:00:00.000Z
#> 3    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#> 4    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#> 5    kato_1993_19910901_948 1991-09-01T04:00:00.000Z
#>                                                                                                                                    networks.description
#> 1                                                                         Insects observed pollinating flowers, ten miles of Carlinville, Illinois, USA
#> 2                                                                         Insects observed pollinating flowers, ten miles of Carlinville, Illinois, USA
#> 3 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#> 4 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#> 5 Flower and anthophilous insect interactions in the primary cool-temperate subalpine forests and meadows at Mt. Kushigata, Yamanashi Prefecture, Japan
#>   networks.public networks.all_interactions      networks.created_at
#> 1            TRUE                     FALSE 2019-02-22T18:47:30.657Z
#> 2            TRUE                     FALSE 2019-02-22T18:47:30.657Z
#> 3            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#> 4            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#> 5            TRUE                     FALSE 2019-02-25T20:52:09.499Z
#>        networks.updated_at networks.dataset_id networks.user_id
#> 1 2019-02-22T18:47:30.657Z                   1                3
#> 2 2019-02-22T18:47:30.657Z                   1                3
#> 3 2019-02-25T20:52:09.499Z                  66                2
#> 4 2019-02-25T20:52:09.499Z                  66                2
#> 5 2019-02-25T20:52:09.499Z                  66                2
#>              networks.geom
#> 1 POINT (-89.8818 39.2798)
#> 2 POINT (-89.8818 39.2798)
#> 3 POINT (138.3833 35.5833)
#> 4 POINT (138.3833 35.5833)
#> 5 POINT (138.3833 35.5833)

# Retrieve higher classification
tsn_acer <- search_taxa("Acer")$taxonomy.tsn
search_taxa(list(network_id = 926), original = TRUE)

#>      id   original_name node_level network_id taxonomy_id
#> 1  6445        mollusks      taxon        926        4336
#> 2  6447         insects      taxon        926        4650
#> 3  6448     Callinectes      taxon        926        3823
#> 4  6449         Penaeus      taxon        926        4806
#> 5  6450 fishes (type 1)      taxon        926        4173
#> 6  6452 fishes (type 3)      taxon        926        4807
#> 7  6453   benthic algae      taxon        926        4309
#> 8  6454        detritus      taxon        926          NA
#> 9  6446     zooplankton      taxon        926          NA
#> 10 6451 fishes (type 2)      taxon        926          NA
#> 11 6455   phytoplankton      taxon        926          NA
#> 12 6443          micro-      taxon        926          NA
#> 13 6444    meio-benthos      taxon        926          NA
#>                  created_at               updated_at taxonomy.id
#> 1  2019-02-25T15:09:24.587Z 2019-02-25T15:09:24.587Z        4336
#> 2  2019-02-25T15:09:24.658Z 2019-02-25T15:09:24.658Z        4650
#> 3  2019-02-25T15:09:24.701Z 2019-02-25T15:09:24.701Z        3823
#> 4  2019-02-25T15:09:24.733Z 2019-02-25T15:09:24.733Z        4806
#> 5  2019-02-25T15:09:24.765Z 2019-02-25T15:09:24.765Z        4173
#> 6  2019-02-25T15:09:24.843Z 2019-02-25T15:09:24.843Z        4807
#> 7  2019-02-25T15:09:24.890Z 2019-02-25T15:09:24.890Z        4309
#> 8  2019-02-25T15:09:24.933Z 2019-02-25T15:09:24.933Z          NA
#> 9  2019-02-25T15:09:24.625Z 2019-02-25T15:09:24.625Z          NA
#> 10 2019-02-25T15:09:24.799Z 2019-02-25T15:09:24.799Z          NA
#> 11 2019-02-25T15:09:24.979Z 2019-02-25T15:09:24.979Z          NA
#> 12 2019-02-25T15:09:24.496Z 2019-02-25T15:09:24.496Z          NA
#> 13 2019-02-25T15:09:24.543Z 2019-02-25T15:09:24.543Z          NA
#>         taxonomy.name taxonomy.ncbi taxonomy.tsn taxonomy.eol taxonomy.bold
#> 1            Mollusca          6447        69458         2195            23
#> 2             Insecta         50557         4703          344            82
#> 3         Callinectes          6762        13951     46508442          4985
#> 4             Penaeus        133894        95603     46502054          5629
#> 5        Salmo trutta          8032       161997       206777         47318
#> 6  Salvelinus alpinus          8036       162001      1156463         80552
#> 7     Archanara algae        131213         9417       522707         22849
#> 8                <NA>            NA           NA           NA            NA
#> 9                <NA>            NA           NA           NA            NA
#> 10               <NA>            NA           NA           NA            NA
#> 11               <NA>            NA           NA           NA            NA
#> 12               <NA>            NA           NA           NA            NA
#> 13               <NA>            NA           NA           NA            NA
#>    taxonomy.gbif                     taxonomy.col taxonomy.rank
#> 1             NA 70a72224d52b24f5ee0b1d09387d1703        phylum
#> 2             NA ae304a1e0beadcfec04932589049bb5a         class
#> 3             NA 8c9d5b7b0ef67ba0ca49a80b06665541         genus
#> 4             NA 854dd2c279c74204cd65501818d105d4         genus
#> 5        8215487 3d9e5032ec6a0c4b846544e2db56fa0a       species
#> 6        4284021 2b256e111056ab998c15402950cf0b1b       species
#> 7             NA f78d2ff8715be72fd77608d38b07cf32       species
#> 8             NA                             <NA>          <NA>
#> 9             NA                             <NA>          <NA>
#> 10            NA                             <NA>          <NA>
#> 11            NA                             <NA>          <NA>
#> 12            NA                             <NA>          <NA>
#> 13            NA                             <NA>          <NA>
#>         taxonomy.created_at      taxonomy.updated_at taxonomy networks.id
#> 1  2019-02-23T14:59:23.089Z 2019-06-14T15:25:44.735Z       NA         926
#> 2  2019-02-24T18:56:42.545Z 2019-06-14T15:25:59.300Z       NA         926
#> 3  2019-02-22T14:55:39.819Z 2019-06-14T15:25:29.810Z       NA         926
#> 4  2019-02-24T23:42:56.274Z 2019-06-14T15:26:06.793Z       NA         926
#> 5  2019-02-22T22:40:06.067Z 2019-06-14T15:25:36.207Z       NA         926
#> 6  2019-02-24T23:42:56.358Z 2019-06-14T15:26:06.915Z       NA         926
#> 7  2019-02-23T14:12:12.304Z 2019-06-14T15:25:43.663Z       NA         926
#> 8                      <NA>                     <NA>       NA         926
#> 9                      <NA>                     <NA>       NA         926
#> 10                     <NA>                     <NA>       NA         926
#> 11                     <NA>                     <NA>       NA         926
#> 12                     <NA>                     <NA>       NA         926
#> 13                     <NA>                     <NA>       NA         926
#>              networks.name            networks.date
#> 1  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 2  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 3  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 4  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 5  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 6  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 7  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 8  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 9  yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 10 yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 11 yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 12 yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#> 13 yanez_1978_19730901_926 1973-09-01T04:00:00.000Z
#>             networks.description networks.public networks.all_interactions
#> 1  Food web of the Costal lagoon            TRUE                     FALSE
#> 2  Food web of the Costal lagoon            TRUE                     FALSE
#> 3  Food web of the Costal lagoon            TRUE                     FALSE
#> 4  Food web of the Costal lagoon            TRUE                     FALSE
#> 5  Food web of the Costal lagoon            TRUE                     FALSE
#> 6  Food web of the Costal lagoon            TRUE                     FALSE
#> 7  Food web of the Costal lagoon            TRUE                     FALSE
#> 8  Food web of the Costal lagoon            TRUE                     FALSE
#> 9  Food web of the Costal lagoon            TRUE                     FALSE
#> 10 Food web of the Costal lagoon            TRUE                     FALSE
#> 11 Food web of the Costal lagoon            TRUE                     FALSE
#> 12 Food web of the Costal lagoon            TRUE                     FALSE
#> 13 Food web of the Costal lagoon            TRUE                     FALSE
#>         networks.created_at      networks.updated_at networks.dataset_id
#> 1  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 2  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 3  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 4  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 5  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 6  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 7  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 8  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 9  2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 10 2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 11 2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 12 2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#> 13 2019-02-25T15:09:10.651Z 2019-02-25T15:09:10.651Z                  52
#>    networks.user_id              networks.geom
#> 1                 3 POINT (-100.3656 17.06915)
#> 2                 3 POINT (-100.3656 17.06915)
#> 3                 3 POINT (-100.3656 17.06915)
#> 4                 3 POINT (-100.3656 17.06915)
#> 5                 3 POINT (-100.3656 17.06915)
#> 6                 3 POINT (-100.3656 17.06915)
#> 7                 3 POINT (-100.3656 17.06915)
#> 8                 3 POINT (-100.3656 17.06915)
#> 9                 3 POINT (-100.3656 17.06915)
#> 10                3 POINT (-100.3656 17.06915)
#> 11                3 POINT (-100.3656 17.06915)
#> 12                3 POINT (-100.3656 17.06915)
#> 13                3 POINT (-100.3656 17.06915)