Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

Search network by taxa using keyword, tsn, eol, bold or ncbi IDs.

search_taxa(query = NULL, original = FALSE, verbose = TRUE, ...)

Arguments

query
character specific taxon / keyword / or tsn, eol, bold & ncbi IDs (mandatory arg).

original
a logical. Should query taxa names from the original publication? (see details section, default is set to FALSE)

verbose
a logical. Should extra information be reported on progress?

...
further arguments to be passed to rmangal::get_gen().

Value

An object of class mgSearchTaxa, which is a data.frame object with all taxa matching the query. All networks in which taxa are involved are also attached to the data.frame.

Details

By default, you are querying taxa based on a homogenized taxonomy (corrected names from TNRS and/or GNR. These taxa might not be the taxon name used in the original publication. In order to query, the original name, set argument original to TRUE.

Examples

search_taxa("Acer")

#> Found 5 taxa

#> # A tibble: 5 x 11
#>      id name    ncbi    tsn    eol   bold gbif  created_at updated_at
#>   <int> <chr>  <int>  <int>  <int>  <int> <lgl> <chr>      <chr>     
#> 1     2 Acer…   4023 182132 5.09e7 100987 NA    2019-02-2… 2019-02-2…
#> 2     3 Acer…  75745  28757 5.73e7 101028 NA    2019-02-2… 2019-02-2…
#> 3  3117 Acer…  66216     NA 5.73e7 485602 NA    2019-02-2… 2019-02-2…
#> 4  3144 Acer… 171221     NA 5.73e7 102197 NA    2019-02-2… 2019-02-2…
#> 5  3151 Acer…  47966 837855 5.73e7 448667 NA    2019-02-2… 2019-02-2…
#> # … with 13 more variables: original_name <chr>, networks$id <int>,
#> #   $name <chr>, $date <chr>, $geom.type <chr>, $geom.coordinates <dbl>,
#> #   $description <chr>, $public <lgl>, $all_interactions <lgl>,
#> #   $created_at <chr>, $updated_at <chr>, $dataset_id <int>, $user_id <int>

# Retrieve higher classification
tsn_acer <- search_taxa("Acer")$tsn

#> Found 5 taxa

taxize::classification(tsn_acer, db = "itis")

#> $`182132`
#>                           name          rank     id
#> 1                      Plantae       kingdom 202422
#> 2                Viridiplantae    subkingdom 954898
#> 3                 Streptophyta  infrakingdom 846494
#> 4                  Embryophyta superdivision 954900
#> 5                 Tracheophyta      division 846496
#> 6              Spermatophytina   subdivision 846504
#> 7                Magnoliopsida         class  18063
#> 8                      Rosanae    superorder 846548
#> 9                   Sapindales         order  28643
#> 10                 Sapindaceae        family  28657
#> 11                        Acer         genus  28727
#> 12                Acer negundo       species  28749
#> 13 Acer negundo var. violaceum       variety 182132
#> 
#> $`28757`
#>                name          rank     id
#> 1           Plantae       kingdom 202422
#> 2     Viridiplantae    subkingdom 954898
#> 3      Streptophyta  infrakingdom 846494
#> 4       Embryophyta superdivision 954900
#> 5      Tracheophyta      division 846496
#> 6   Spermatophytina   subdivision 846504
#> 7     Magnoliopsida         class  18063
#> 8           Rosanae    superorder 846548
#> 9        Sapindales         order  28643
#> 10      Sapindaceae        family  28657
#> 11             Acer         genus  28727
#> 12 Acer saccharinum       species  28757
#> 
#> $<NA>
#> [1] NA
#> 
#> $<NA>
#> [1] NA
#> 
#> $`837855`
#>               name          rank     id
#> 1          Plantae       kingdom 202422
#> 2    Viridiplantae    subkingdom 954898
#> 3     Streptophyta  infrakingdom 846494
#> 4      Embryophyta superdivision 954900
#> 5     Tracheophyta      division 846496
#> 6  Spermatophytina   subdivision 846504
#> 7    Magnoliopsida         class  18063
#> 8          Rosanae    superorder 846548
#> 9       Sapindales         order  28643
#> 10     Sapindaceae        family  28657
#> 11            Acer         genus  28727
#> 12  Acer japonicum       species 837855
#> 
#> attr(,"class")
#> [1] "classification"
#> attr(,"db")
#> [1] "itis"